requirements.txt
uvloop==0.12.2
uvicorn
starlette
google-cloud-storage